requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
watchfiles==0.21.0
